Finished the seperation
